@jreback 2020-07-06
very nice, cleaning up some nasty code + perf!