@jbrockmendel 2020-07-06
> cleaning up some nasty code 

Yep.  #35108, #35106, #35104, and #35102 together fix our basically-incorrect tz_convert_single usage, at which point we can clean up a bunch more.