@jreback 2020-07-07
i merged everything above (i think)
pls give a check